ceres-protocol/services/ai-agent/pyproject.toml
[tool.poetry]
name = "ceres-ai-agent"
version = "0.1.0"
description = "AI Intelligent Agent for Ceres Protocol - Multi-mode prediction market agent"
authors = ["Ceres Protocol Team"]
readme = "README.md"
packages = [{include = "ai_agent"}]

[tool.poetry.dependencies]
python = "^3.11"
web3 = "^6.15.1"
asyncio = "^3.4.3"
pydantic = "^2.5.3"
requests = "^2.31.0"
numpy = "^1.26.3"
python-dotenv = "^1.0.0"
aiohttp = "^3.9.1"
eth-account = "^0.10.0"
eth-typing = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.2"
hypothesis = "^6.92.1"
black = "^23.12.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]